+++ package.json
@@ -58,5 +58,6 @@
     "@parcel/watcher": "^2.5.1",
     "graphql": "^16",
     "typescript": "^5"
-  }
+  },
+  "packageManager": "yarn@3.5.0+sha512.2dc70be5fce9f66756d25b00a888f3ca66f86b502b76750e72ba54cec89da767b938c54124595e26f868825688e0fe3552c26c76a330673343057acadd5cfcf2"
 }
